Add new static method to generate basis functions Lobster
