+++ pymatgen/io/lobster/__init__.py
@@ -1,0 +1,14 @@
+# coding: utf-8
+# Copyright (c) Pymatgen Development Team.
+# Distributed under the terms of the MIT License.
+
+
+"""
+This package implements modules for input and output to and from Lobster. It
+imports the key classes form both lobster.inputs and lobster_outputs to allow most
+classes to be simply called as pymatgen.io.lobster.Lobsterin for example, to retain
+backwards compatibility.
+"""
+
+from .inputs import Lobsterin
+from .outputs import Cohpcar,  Icohplist, Doscar, Charge, Lobsterout, Fatband, Bandoverlaps, Grosspop
