+++ pymatgen/io/lobster/lobster_basis/BASIS_PBE_54_min.yaml
@@ -1,0 +1,189 @@
+BASIS:
+  Ac: '6d 6p 6s 7s '
+  Ag: '4d 5s '
+  Ag_pv: '4d 4p 5s '
+  Al: '3p 3s '
+  Am: '5f 6d 6p 6s 7s '
+  Ar: '3p 3s '
+  As: '4p 4s '
+  As_d: '3d 4p 4s '
+  At: '6p 6s '
+  Au: '5d 6s '
+  B: '2p 2s '
+  B_h: '2p 2s '
+  B_s: '2p 2s '
+  Ba_sv: '5p 5s 6s '
+  Be: '2p 2s '
+  Be_sv: '1s 2p 2s '
+  Bi: '6p 6s '
+  Bi_d: '5d 6p 6s '
+  Br: '4p 4s '
+  C: '2p 2s '
+  C_h: '2p 2s '
+  C_s: '2p 2s '
+  Ca_pv: '3p 4s '
+  Ca_sv: '3p 3s 4s '
+  Cd: '4d 5s '
+  Ce: '4f 5d 5p 5s 6s '
+  Ce_3: '5d 5p 5s 6s '
+  Ce_h: '4f 5d 5p 5s 6s '
+  Cf: '5f 6p 6s 7s '
+  Cl: '3p 3s '
+  Cl_h: '3p 3s '
+  Cm: '5f 6d 6p 6s 7s '
+  Co: '3d 4s '
+  Co_pv: '3d 3p 4s '
+  Co_sv: '3d 3p 3s 4s '
+  Cr: '3d 4s '
+  Cr_pv: '3d 3p 4s '
+  Cr_sv: '3d 3p 3s 4s '
+  Cs_sv: '5p 5s 6s '
+  Cu: '3d 4s '
+  Cu_pv: '3d 3p 4s '
+  Dy: '4f 5d 5p 5s 6s '
+  Dy_3: '5d 5p 6s '
+  Er: '4f 5d 5p 5s 6s '
+  Er_2: '5p 6s '
+  Er_3: '5d 5p 6s '
+  Eu: '4f 5p 5s 6s '
+  Eu_2: '5p 6s '
+  Eu_3: '5d 5p 6s '
+  F: '2p 2s '
+  F_h: '2p 2s '
+  F_s: '2p 2s '
+  Fe: '3d 4s '
+  Fe_pv: '3d 3p 4s '
+  Fe_sv: '3d 3p 3s 4s '
+  Fr_sv: '6p 6s 7s '
+  Ga: '4p 4s '
+  Ga_d: '3d 4p 4s '
+  Ga_h: '3d 4p 4s '
+  Gd: '4f 5d 5p 5s 6s '
+  Gd_3: '5d 5p 6s '
+  Ge: '4p 4s '
+  Ge_d: '3d 4p 4s '
+  Ge_h: '3d 4p 4s '
+  H: '1s '
+  H_h: '1s '
+  H_s: '1s '
+  He: '1s '
+  Hf: '5d 6s '
+  Hf_pv: '5d 5p 6s '
+  Hf_sv: '5d 5p 5s '
+  Hg: '5d 6s '
+  Ho: '4f 5d 5p 5s 6s '
+  Ho_3: '5d 5p 6s '
+  I: '5p 5s '
+  In: '5p 5s '
+  In_d: '4d 5p 5s '
+  Ir: '5d 6s '
+  K_pv: '3p 4s '
+  K_sv: '3p 3s 4s '
+  Kr: '4p 4s '
+  La: '5d 5p 5s 6s '
+  La_s: '5d 5p 6s '
+  Li: '2s '
+  Li_sv: '1s 2s '
+  Lu: '4f 5d 5p 5s 6s '
+  Lu_3: '5d 5p 6s '
+  Mg: '3s '
+  Mg_pv: '2p 3s '
+  Mg_sv: '2p 2s 3s '
+  Mn: '3d 4s '
+  Mn_pv: '3d 3p 4s '
+  Mn_sv: '3d 3p 3s 4s '
+  Mo: '4d 5s '
+  Mo_pv: '4d 4p 5s '
+  Mo_sv: '4d 4p 4s 5s '
+  N: '2p 2s '
+  N_h: '2p 2s '
+  N_s: '2p 2s '
+  Na: '3s '
+  Na_pv: '2p 3s '
+  Na_sv: '2p 2s 3s '
+  Nb_pv: '4d 4p 5s '
+  Nb_sv: '4d 4p 4s 5s '
+  Nd: '4f 5d 5p 5s 6s '
+  Nd_3: '5d 5p 5s 6s '
+  Ne: '2p 2s '
+  Ni: '3d 4s '
+  Ni_pv: '3d 3p 4s '
+  Np: '5f 6d 6p 6s 7s '
+  Np_s: '5f 6d 6p 6s 7s '
+  O: '2p 2s '
+  O_h: '2p 2s '
+  O_s: '2p 2s '
+  Os: '5d 6s '
+  Os_pv: '5d 5p 6s '
+  P: '3p 3s '
+  P_h: '3p 3s '
+  Pa: '5f 6d 6p 6s 7s '
+  Pa_s: '5f 6d 6p 7s '
+  Pb: '6p 6s '
+  Pb_d: '5d 6p 6s '
+  Pd: '4d 5s '
+  Pd_pv: '4d 4p 5s '
+  Pm: '4f 5d 5p 5s 6s '
+  Pm_3: '5d 5p 5s 6s '
+  Po: '6p 6s '
+  Po_d: '5d 6p 6s '
+  Pr: '4f 5d 5p 5s 6s '
+  Pr_3: '5d 5p 5s 6s '
+  Pt: '5d 6s '
+  Pt_pv: '5d 5p 6s '
+  Pu: '5f 6d 6p 6s 7s '
+  Pu_s: '5f 6d 6p 6s 7s '
+  Ra_sv: '6p 6s 7s '
+  Rb_pv: '4p 5s '
+  Rb_sv: '4p 4s 5s '
+  Re: '5d 6s '
+  Re_pv: '5d 5p 6s '
+  Rh: '4d 5s '
+  Rh_pv: '4d 4p 5s '
+  Rn: '6p 6s '
+  Ru: '4d 5s '
+  Ru_pv: '4d 4p 5s '
+  Ru_sv: '4d 4p 4s 5s '
+  S: '3p 3s '
+  S_h: '3p 3s '
+  Sb: '5p 5s '
+  Sc: '3d 4s '
+  Sc_sv: '3d 3p 3s 4s '
+  Se: '4p 4s '
+  Si: '3p 3s '
+  Sm: '4f 5d 5p 5s 6s '
+  Sm_3: '5d 5p 5s 6s '
+  Sn: '5p 5s '
+  Sn_d: '4d 5p 5s '
+  Sr_sv: '4p 4s 5s '
+  Ta: '5d 6s '
+  Ta_pv: '5d 5p 6s '
+  Tb: '4f 5d 5p 5s 6s '
+  Tb_3: '5d 5p 6s '
+  Tc: '4d 5s '
+  Tc_pv: '4d 4p 5s '
+  Tc_sv: '4d 4p 4s 5s '
+  Te: '5p 5s '
+  Th: '5f 6d 6p 6s 7s '
+  Th_s: '5f 6d 6p 7s '
+  Ti: '3d 4s '
+  Ti_pv: '3d 3p 4s '
+  Ti_sv: '3d 3p 3s 4s '
+  Tl: '6p 6s '
+  Tl_d: '5d 6p 6s '
+  Tm: '4f 5d 5p 5s 6s '
+  Tm_3: '5d 5p 6s '
+  U: '5f 6d 6p 6s 7s '
+  U_s: '5f 6d 6p 6s 7s '
+  V: '3d 4s '
+  V_pv: '3d 3p 4s '
+  V_sv: '3d 3p 3s 4s '
+  W: '5d 6s '
+  W_sv: '5d 5p 5s 6s '
+  Xe: '5p 5s '
+  Y_sv: '4d 4p 4s 5s '
+  Yb: '4f 5p 5s 6s '
+  Yb_2: '5p 6s '
+  Yb_3: '5d 5p 6s '
+  Zn: '3d 4s '
+  Zr_sv: '4d 4p 4s 5s '
