+++ pymatgen/io/lobster/tests/test_lobster.py
@@ -9,10 +9,12 @@ import unittest
 import warnings
 
 import numpy as np
+
 from pymatgen import Structure
 from pymatgen.electronic_structure.core import Spin, Orbital
 from pymatgen.io.lobster import Cohpcar, Icohplist, Doscar, Charge, Lobsterout, Fatband, Lobsterin, Bandoverlaps, \
     Grosspop
+from pymatgen.io.lobster.inputs import get_all_possible_basis_combinations
 from pymatgen.io.vasp import Vasprun
 from pymatgen.io.vasp.inputs import Incar, Kpoints, Potcar
 from pymatgen.util.testing import PymatgenTest
@@ -24,9 +26,9 @@ __email__ = "janine.george@uclouvain.be, esters@uoregon.edu"
 __date__ = "Dec 10, 2017"
 
 test_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)),
-                        "..", "..", "..", "test_files", "cohp")
+                        "..", "..", "..", "..", "test_files", "cohp")
 test_dir_doscar = os.path.join(os.path.dirname(os.path.abspath(__file__)),
-                               "..", "..", "..", "test_files")
+                               "..", "..", "..", "..", "test_files")
 
 this_dir = os.path.dirname(os.path.abspath(__file__))
 
@@ -361,7 +363,9 @@ class DoscarTest(unittest.TestCase):
         self.assertListEqual(tdos_up, self.DOSCAR_spin_pol.completedos.densities[Spin.up].tolist())
         self.assertListEqual(tdos_down, self.DOSCAR_spin_pol.completedos.densities[Spin.down].tolist())
         self.assertAlmostEqual(fermi, self.DOSCAR_spin_pol.completedos.efermi)
-        self.assertDictEqual(self.DOSCAR_spin_pol.completedos.structure.as_dict(), self.structure.as_dict())
+        for coords, coords2 in zip(self.DOSCAR_spin_pol.completedos.structure.frac_coords, self.structure.frac_coords):
+            for xyz, xyz2 in zip(coords, coords2):
+                self.assertAlmostEqual(xyz, xyz2)
         self.assertListEqual(self.DOSCAR_spin_pol.completedos.pdos[self.structure[0]]['2s'][Spin.up].tolist(),
                              PDOS_F_2s_up)
         self.assertListEqual(self.DOSCAR_spin_pol.completedos.pdos[self.structure[0]]['2s'][Spin.down].tolist(),
@@ -535,7 +539,7 @@ class ChargeTest(PymatgenTest):
                                                              'species': [{'occu': 1, 'element': 'Mn'}], 'label': 'Mn'}],
                            'charge': None, '@module': 'pymatgen.core.structure'}
         s2 = Structure.from_dict(structure_dict2)
-        self.assertEqual(s2, self.charge2.get_structure_with_charges(os.path.join(this_dir, "POSCAR.MnO")))
+        self.assertEqual(s2, self.charge2.get_structure_with_charges(os.path.join(this_dir, "../../tests/POSCAR.MnO")))
 
 
 class LobsteroutTest(PymatgenTest):
@@ -1223,14 +1227,30 @@ class LobsterinTest(unittest.TestCase):
         potcar = Potcar.from_file(os.path.join(test_dir_doscar, "POTCAR.Fe3O4"))
         Potcar_names = [name["symbol"] for name in potcar.spec]
 
-        self.assertListEqual(lobsterin1._get_basis(Structure.from_file(os.path.join(test_dir_doscar, "Fe3O4.cif")),
-                                                   potcar_symbols=Potcar_names),
+        self.assertListEqual(lobsterin1.get_basis(Structure.from_file(os.path.join(test_dir_doscar, "Fe3O4.cif")),
+                                                  potcar_symbols=Potcar_names),
                              ['Fe 3d 4p 4s ', 'O 2p 2s '])
         potcar = Potcar.from_file(os.path.join(test_dir, "POTCAR.GaAs"))
         Potcar_names = [name["symbol"] for name in potcar.spec]
-        self.assertListEqual(lobsterin1._get_basis(Structure.from_file(os.path.join(test_dir, "POSCAR.GaAs")),
-                                                   potcar_symbols=Potcar_names),
+        self.assertListEqual(lobsterin1.get_basis(Structure.from_file(os.path.join(test_dir, "POSCAR.GaAs")),
+                                                  potcar_symbols=Potcar_names),
                              ['Ga 3d 4p 4s ', 'As 4p 4s '])
+
+    def test_get_all_possible_basis_functions(self):
+        potcar = Potcar.from_file(os.path.join(test_dir_doscar, "POTCAR.Fe3O4"))
+        Potcar_names = [name["symbol"] for name in potcar.spec]
+        result = Lobsterin.get_all_possible_basis_functions(
+            Structure.from_file(os.path.join(test_dir_doscar, "Fe3O4.cif")),
+            potcar_symbols=Potcar_names)
+        self.assertDictEqual(result[0], {'Fe': '3d 4s', 'O': '2p 2s'})
+        self.assertDictEqual(result[1], {'Fe': '3d 4s 4p', 'O': '2p 2s'})
+
+        potcar2 = Potcar.from_file(os.path.join(test_dir_doscar, "POT_GGA_PAW_PBE_54/POTCAR.Fe_pv.gz"))
+        Potcar_names2 = [name["symbol"] for name in potcar2.spec]
+        result2 = Lobsterin.get_all_possible_basis_functions(
+            Structure.from_file(os.path.join(test_dir_doscar, "Fe.cif")),
+            potcar_symbols=Potcar_names2)
+        self.assertDictEqual(result2, {'Fe': '3d 3p 4s'})
 
     def test_get_potcar_symbols(self):
         lobsterin1 = Lobsterin({})
@@ -1481,8 +1501,6 @@ class GrosspopTest(unittest.TestCase):
         self.assertAlmostEqual(self.grosspop1.list_dict_grosspop[8]["Loewdin GP"]["2s"], 1.60)
         self.assertEqual(self.grosspop1.list_dict_grosspop[8]["element"], 'O')
 
-    @unittest.skipIf(True, "This test is just skipped for now. One really shouldn't write fragile tests that compare "
-                           "dicts of float for equality")
     def test_structure_with_grosspop(self):
         struct_dict = {'@module': 'pymatgen.core.structure', '@class': 'Structure', 'charge': None, 'lattice': {
             'matrix': [[5.021897888834907, 4.53806e-11, 0.0], [-2.5109484443388332, 4.349090983701526, 0.0],
@@ -1521,9 +1539,50 @@ class GrosspopTest(unittest.TestCase):
              'abc': [0.841096220194068, 0.5850993824105675, 0.8719852747855317],
              'xyz': [2.754744948452184, 2.5446504486493, 4.806321279926453], 'label': 'O',
              'properties': {'Total Mulliken GP': 7.18, 'Total Loewdin GP': 6.92}}]}
-        newstructure = self.grosspop1.get_structure_with_total_grosspop(os.path.join(test_dir, "POSCAR.SiO2"))
 
-        self.assertDictEqual(newstructure.as_dict(), struct_dict)
+        newstructure = self.grosspop1.get_structure_with_total_grosspop(os.path.join(test_dir, "POSCAR.SiO2"))
+        for coords, coords2 in zip(newstructure.frac_coords, Structure.from_dict(struct_dict).frac_coords):
+            for xyz, xyz2 in zip(coords, coords2):
+                self.assertAlmostEqual(xyz, xyz2)
+
+
+class TestUtils(PymatgenTest):
+
+    def test_get_all_possible_basis_combinations(self):
+        # this basis is just for testing (not correct)
+        min_basis = ['Li 1s 2s ', 'Na 1s 2s', 'Si 1s 2s']
+        max_basis = ['Li 1s 2p 2s ', 'Na 1s 2p 2s', 'Si 1s 2s']
+        combinations_basis = get_all_possible_basis_combinations(min_basis, max_basis)
+        self.assertListEqual(combinations_basis,
+                             [['Li 1s 2s', 'Na 1s 2s', 'Si 1s 2s'], ['Li 1s 2s', 'Na 1s 2s 2p', 'Si 1s 2s'],
+                              ['Li 1s 2s 2p', 'Na 1s 2s', 'Si 1s 2s'], ['Li 1s 2s 2p', 'Na 1s 2s 2p', 'Si 1s 2s']])
+
+        min_basis = ['Li 1s 2s']
+        max_basis = ['Li 1s 2s 2p 3s']
+        combinations_basis = get_all_possible_basis_combinations(min_basis, max_basis)
+        self.assertListEqual(combinations_basis, [['Li 1s 2s'], ['Li 1s 2s 2p'], ['Li 1s 2s 3s'], ['Li 1s 2s 2p 3s']])
+
+        min_basis = ['Li 1s 2s', 'Na 1s 2s']
+        max_basis = ['Li 1s 2s 2p 3s', 'Na 1s 2s 2p 3s']
+        combinations_basis = get_all_possible_basis_combinations(min_basis, max_basis)
+        self.assertListEqual(combinations_basis,
+                             [['Li 1s 2s', 'Na 1s 2s'], ['Li 1s 2s', 'Na 1s 2s 2p'], ['Li 1s 2s', 'Na 1s 2s 3s'],
+                              ['Li 1s 2s', 'Na 1s 2s 2p 3s'], ['Li 1s 2s 2p', 'Na 1s 2s'],
+                              ['Li 1s 2s 2p', 'Na 1s 2s 2p'], ['Li 1s 2s 2p', 'Na 1s 2s 3s'],
+                              ['Li 1s 2s 2p', 'Na 1s 2s 2p 3s'], ['Li 1s 2s 3s', 'Na 1s 2s'],
+                              ['Li 1s 2s 3s', 'Na 1s 2s 2p'], ['Li 1s 2s 3s', 'Na 1s 2s 3s'],
+                              ['Li 1s 2s 3s', 'Na 1s 2s 2p 3s'], ['Li 1s 2s 2p 3s', 'Na 1s 2s'],
+                              ['Li 1s 2s 2p 3s', 'Na 1s 2s 2p'], ['Li 1s 2s 2p 3s', 'Na 1s 2s 3s'],
+                              ['Li 1s 2s 2p 3s', 'Na 1s 2s 2p 3s']])
+
+        min_basis = ['Si 1s 2s 2p', 'Na 1s 2s']
+        max_basis = ['Si 1s 2s 2p 3s', 'Na 1s 2s 2p 3s']
+        combinations_basis = get_all_possible_basis_combinations(min_basis, max_basis)
+        self.assertListEqual(combinations_basis, [['Si 1s 2s 2p', 'Na 1s 2s'], ['Si 1s 2s 2p', 'Na 1s 2s 2p'],
+                                                  ['Si 1s 2s 2p', 'Na 1s 2s 3s'], ['Si 1s 2s 2p', 'Na 1s 2s 2p 3s'],
+                                                  ['Si 1s 2s 2p 3s', 'Na 1s 2s'], ['Si 1s 2s 2p 3s', 'Na 1s 2s 2p'],
+                                                  ['Si 1s 2s 2p 3s', 'Na 1s 2s 3s'],
+                                                  ['Si 1s 2s 2p 3s', 'Na 1s 2s 2p 3s']])
 
 
 if __name__ == "__main__":
